Move declarations into its own namespace
