+++ packages/analyzer/src/Declarations/Class_Property.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Automattic\Jetpack\Analyzer;
+namespace Automattic\Jetpack\Analyzer\Declarations;
 
 /**
  * We only log public class variables
  */
-class Class_Property_Declaration extends Declaration {
+class Class_Property extends Declaration {
 	public $class_name;
 	public $name;
 	public $static;
