+++ packages/analyzer/src/Declarations/Visitor.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Automattic\Jetpack\Analyzer;
+namespace Automattic\Jetpack\Analyzer\Declarations;
 
 use PhpParser\NodeVisitorAbstract;
 use PhpParser\Node;
 
-class Declaration_Visitor extends NodeVisitorAbstract {
+class Visitor extends NodeVisitorAbstract {
 	private $current_class;
 	private $declarations;
 	private $current_relative_path;
@@ -20,10 +20,10 @@ class Declaration_Visitor extends NodeVisitorAbstract {
 			// $this->current_class = $node->name->name;
 			$this->current_class = implode( '\\', $node->namespacedName->parts );
 
-			$this->declarations->add( new Class_Declaration( $this->current_relative_path, $node->getLine(), $node->name->name ) );
+			$this->declarations->add( new Class_( $this->current_relative_path, $node->getLine(), $node->name->name ) );
 		}
 		if ( $node instanceof Node\Stmt\Property && $node->isPublic() ) {
-			$this->declarations->add( new Class_Property_Declaration( $this->current_relative_path, $node->getLine(), $this->current_class, $node->props[0]->name->name, $node->isStatic() ) );
+			$this->declarations->add( new Class_Property( $this->current_relative_path, $node->getLine(), $this->current_class, $node->props[0]->name->name, $node->isStatic() ) );
 		}
 		if ( $node instanceof Node\Stmt\ClassMethod && $node->isPublic() ) {
 			// ClassMethods are also listed inside interfaces, which means current_class is null
@@ -31,14 +31,14 @@ class Declaration_Visitor extends NodeVisitorAbstract {
 			if ( ! $this->current_class ) {
 				return;
 			}
-			$method = new Class_Method_Declaration( $this->current_relative_path, $node->getLine(), $this->current_class, $node->name->name, $node->isStatic() );
+			$method = new Class_Method( $this->current_relative_path, $node->getLine(), $this->current_class, $node->name->name, $node->isStatic() );
 			foreach ( $node->getParams() as $param ) {
 				$method->add_param( $param->var->name, $param->default, $param->type, $param->byRef, $param->variadic );
 			}
 			$this->declarations->add( $method );
 		}
 		if ( $node instanceof Node\Stmt\Function_ ) {
-			$function = new Function_Declaration( $this->current_relative_path, $node->getLine(), $node->name->name );
+			$function = new Function_( $this->current_relative_path, $node->getLine(), $node->name->name );
 			foreach ( $node->getParams() as $param ) {
 				$function->add_param( $param->var->name, $param->default, $param->type, $param->byRef, $param->variadic );
 			}
