+++ packages/analyzer/src/Invocations/Invocation_Visitor.php
@@ -16,10 +16,10 @@ class Invocation_Visitor extends NodeVisitorAbstract {
 
 		// if ( $node instanceof Node\Stmt\Class_ ) {
 		// 	$this->current_class = $node->name->name;
-		// 	$this->add( new Class_Declaration( $this->current_relative_path, $node->getLine(), $node->name->name ) );
+		// 	$this->add( new Declarations\Class_( $this->current_relative_path, $node->getLine(), $node->name->name ) );
 		// }
 		// if ( $node instanceof Node\Stmt\Property && $node->isPublic() ) {
-		// 	$this->add( new Class_Property_Declaration( $this->current_relative_path, $node->getLine(), $this->current_class, $node->props[0]->name->name, $node->isStatic() ) );
+		// 	$this->add( new Declarations\Class_Property( $this->current_relative_path, $node->getLine(), $this->current_class, $node->props[0]->name->name, $node->isStatic() ) );
 		// }
 		// if ( $node instanceof Node\Stmt\ClassMethod && $node->isPublic() ) {
 		// 	// ClassMethods are also listed inside interfaces, which means current_class is null
@@ -27,7 +27,7 @@ class Invocation_Visitor extends NodeVisitorAbstract {
 		// 	if ( ! $this->current_class ) {
 		// 		return;
 		// 	}
-		// 	$method = new Class_Method_Declaration( $this->current_relative_path, $node->getLine(), $this->current_class, $node->name->name, $node->isStatic() );
+		// 	$method = new Declarations\Class_Method( $this->current_relative_path, $node->getLine(), $this->current_class, $node->name->name, $node->isStatic() );
 		// 	foreach ( $node->getParams() as $param ) {
 		// 		$method->add_param( $param->var->name, $param->default, $param->type, $param->byRef, $param->variadic );
 		// 	}
